Introduce async request for beesmasActive
This is very important if we continue to use GitHub to store other data (currently only beesmasActive and contributors list). The previous method (async=0) increased loading time as NM waited for the HTTP request to finish before continuing with creating the GUI.

The new method simply defaults to the "failed" beesmasActive=0 display, and updates it accordingly once the HTTP request is finished, allowing GUI creation to continue uninterrupted.

This required some other changes in the code, in line order:
- 'hotbarwhilelist' may not be updated  before the async request is done, so I changed this to always show Snowflake but only allow it when beesmasActive=1
- added 'BeesmasFailImage' to Collect/Kill show/hide
- added back 'beesmasActive' as a condition for several interrupts and nm_Collect()